+++ src/main/java/org/scijava/plot/specification/encoding/EncodingStack.java
@@ -13,50 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scijava.plot.spec;
+package org.scijava.plot.specification.encoding;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
  *
  * @author hadim
  */
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
-public class VegaTransform {
-
-    @JsonProperty
-    private String bin;
-
-    @JsonProperty
-    private String field;
-
-    @JsonProperty
-    private String as;
-
-    public String getBin() {
-        return bin;
-    }
-
-    public void setBin(String bin) {
-        this.bin = bin;
-    }
-
-    public String getField() {
-        return field;
-    }
-
-    public void setField(String field) {
-        this.field = field;
-    }
-
-    public String getAs() {
-        return as;
-    }
-
-    public void setAs(String as) {
-        this.as = as;
-    }
-    
+public class EncodingStack {
     
 }
