+++ src/main/java/org/scijava/plot/specification/encoding/EncodingTimeUnit.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scijava.plot.spec.encoding;
+package org.scijava.plot.specification.encoding;
 
 import java.util.Arrays;
